+++ src/main/java/com/localhub/paymentservice/infrastructure/gateway/PaymentGatewayAdapter.java
@@ -3,6 +3,7 @@ package com.localhub.paymentservice.infrastructure.gateway;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -13,22 +14,22 @@ import java.util.UUID;
 @Slf4j
 public class PaymentGatewayAdapter {
 
-    public boolean reserve(Double amount, String paymentMethodId) {
+    public boolean reserve(BigDecimal amount, String paymentMethodId) {
         log.info("Mock gateway: reserving amount={} with paymentMethodId={}", amount, paymentMethodId);
         return true;
     }
 
-    public boolean capture(UUID paymentId, Double amount) {
+    public boolean capture(UUID paymentId, BigDecimal amount) {
         log.info("Mock gateway: capturing paymentId={} amount={}", paymentId, amount);
         return true;
     }
 
-    public boolean release(UUID paymentId, Double amount) {
+    public boolean release(UUID paymentId, BigDecimal amount) {
         log.info("Mock gateway: releasing paymentId={} amount={}", paymentId, amount);
         return true;
     }
 
-    public boolean refund(UUID paymentId, Double amount) {
+    public boolean refund(UUID paymentId, BigDecimal amount) {
         log.info("Mock gateway: refunding paymentId={} amount={}", paymentId, amount);
         return true;
     }
